+++ nokhwa-core/src/simd/yuyv_extract_luma.rs
@@ -41,7 +41,7 @@ pub(crate) fn yuyv_extract_luma_simd(src: &[u8], dst: &mut [u8]) {
 }
 
 #[allow(dead_code)]
-fn yuyv_extract_luma_scalar(src: &[u8], dst: &mut [u8]) {
+pub(crate) fn yuyv_extract_luma_scalar(src: &[u8], dst: &mut [u8]) {
     for (chunk, out) in src.chunks_exact(4).zip(dst.chunks_exact_mut(2)) {
         out[0] = chunk[0]; // Y0
         out[1] = chunk[2]; // Y1